+++ rust/crd/src/constants.rs
@@ -1,3 +1,7 @@
+pub const DEFAULT_DFS_REPLICATION_FACTOR: u8 = 3;
+
+pub const CONTROLLER_NAME: &str = "hdfsclusters.hdfs.stackable.tech";
+
 pub const FIELD_MANAGER_SCOPE: &str = "hdfscluster";
 pub const FIELD_MANAGER_SCOPE_POD: &str = "pod-service";